Add an allow-all profile (#1207)

+++ lib/clientv3/profile.go
@@ -16,9 +16,12 @@ package clientv3
 
 import (
 	"context"
+	"fmt"
 
 	apiv3 "github.com/projectcalico/libcalico-go/lib/apis/v3"
+	cerrors "github.com/projectcalico/libcalico-go/lib/errors"
 	"github.com/projectcalico/libcalico-go/lib/options"
+	cresources "github.com/projectcalico/libcalico-go/lib/resources"
 	validator "github.com/projectcalico/libcalico-go/lib/validator/v3"
 	"github.com/projectcalico/libcalico-go/lib/watch"
 )
@@ -41,6 +44,13 @@ type profiles struct {
 // Create takes the representation of a Profile and creates it.  Returns the stored
 // representation of the Profile, and an error, if there is any.
 func (r profiles) Create(ctx context.Context, res *apiv3.Profile, opts options.SetOptions) (*apiv3.Profile, error) {
+	if res.Name == cresources.DefaultAllowProfileName {
+		return nil, cerrors.ErrorResourceAlreadyExists{
+			Err:        fmt.Errorf("%v already exists. The profile %q is default provided by Calico", cresources.DefaultAllowProfileName, cresources.DefaultAllowProfileName),
+			Identifier: cresources.DefaultAllowProfileName,
+		}
+	}
+
 	if err := validator.Validate(res); err != nil {
 		return nil, err
 	}
@@ -55,6 +65,14 @@ func (r profiles) Create(ctx context.Context, res *apiv3.Profile, opts options.S
 // Update takes the representation of a Profile and updates it. Returns the stored
 // representation of the Profile, and an error, if there is any.
 func (r profiles) Update(ctx context.Context, res *apiv3.Profile, opts options.SetOptions) (*apiv3.Profile, error) {
+	if res.Name == cresources.DefaultAllowProfileName {
+		return nil, cerrors.ErrorOperationNotSupported{
+			Operation:  "Update",
+			Identifier: cresources.DefaultAllowProfileName,
+			Reason:     fmt.Sprintf("The profile %q is a default provided by Calico and cannot be updated", cresources.DefaultAllowProfileName),
+		}
+	}
+
 	if err := validator.Validate(res); err != nil {
 		return nil, err
 	}
@@ -68,6 +86,13 @@ func (r profiles) Update(ctx context.Context, res *apiv3.Profile, opts options.S
 
 // Delete takes name of the Profile and deletes it. Returns an error if one occurs.
 func (r profiles) Delete(ctx context.Context, name string, opts options.DeleteOptions) (*apiv3.Profile, error) {
+	if name == cresources.DefaultAllowProfileName {
+		return nil, cerrors.ErrorOperationNotSupported{
+			Operation:  "Delete",
+			Identifier: cresources.DefaultAllowProfileName,
+			Reason:     fmt.Sprintf("The profile %q is a default provided by Calico and cannot be deleted", cresources.DefaultAllowProfileName),
+		}
+	}
 	out, err := r.client.resources.Delete(ctx, opts, apiv3.KindProfile, noNamespace, name)
 	if out != nil {
 		return out.(*apiv3.Profile), err
